backend/requirements.txt
Django>=4.0
djangorestframework>=3.13
django-cors-headers>=3.11
